Remove totally false sentence
This statement is just not true at all:

"In the case of Send As permissions, if the user and the resource attempting to be sent as aren't moved at the same time, you'll need to explicitly add the Send As permission in Exchange Online using the Add-RecipientPermission cmdlet."

I *think* the intent of that statement may have been to draw attention to the fact that on-premises' ***-ADPermission cmdlets are not available in EXO, rather are replaced by the EXO-only ***-RecipientPermission cmdlets.  But the added fluff about whether or not the Send-As permission holders are migrated at the same time as the mailbox with those permissions granted out, is completely false. 

If I'm wrong, this Pull Request still stands. If there is something true about that point, it should be explained more clearly than that.